[FEATURE] Flesh out user facing API in templates
The context view helpers are now exposing the
context to their children, so that we can get
rid of the ugly "frame.withContext" view helper
and do not need to pass in the context as argument
in inline Fluid syntax any more.

As an additional win, it is now also possible to output
content from these contexts without turbo frames.