[BUGFIX] Revert to put context into data attribute
Having the context in the frame id was a hack,
that bit us, by not being able to properly reference
frames any more.
Now the context is in a data attribute again,
and by doing so, the src building for async loading
must include the context arguments in the URI.
While this does not look too nice, it also
makes it possible to keep the  the async loading
of the Hotwire lib removed, which we needed to do,
to allow to inject the context header from the data attribute
(which was what we wanted to get rid of with putting
the whole context into the frame id).