[FEATURE] Explicit response wrap and action rendering
Magic detection did not work out, made the code
more complex and the behavior hard to understand.
Therefore go with letting users explicitly set this as argument
for a frame.

Also provide the possibility to let slots render a specific
action of a given plugin and also explicitly pass the information
whether controllers should only render the requested frame.