[FEATURE] Allow frames without context
To now benefit from the partial rendering of Topwire,
frames must be wrapped in context view helpers,
otherwise TYPO3 keeps rendering the full page.

To make this work a bit more reliable,
the possibility to render partials has been removed
and the partial rendering of templates within a plugin
is now based on Fluid sections.