[BUGFIX] Do not perform a pid check for pages
One could argue, that it never makes sense to check
whether a parent page is visible, when rendering (parts of)
a page. For sure it does not make sense for Topwire,
when a page is the context record.

When the context record is a page and itself is available,
the rendering should be performed.

This is especially important, when the context record is
the root page, but also for all other cases where the
parent page might be hidden, but the context record page is not.