[FEATURE] Allow topwire context to be build for pageId
The topwire context can be build for different types like
plugins via the page link builder but it was not possible
to provide a pageId where the context is realted to.
So it was not possible to create a topwire link that
points to a different page via the page link builder.