[BUGFIX] Do not allow to override context page id
Instead of needing to pass the context page id
to a link that is created, use the page id of the
page id of the link target.

Otherwise it would be possible to create invalid links,
when link target page id does not match context page id.

For that to work, the query param modification is moved
to a new event listener.