[BUGFIX] Store context and page arguments for USER_INT plugins
When using the context render view helper with an action
argument and the action is uncached, then TYPO3 does not hand
over the correct/modified request, but the original server request.

For that with the current state of TYPO3, the USER content object
class needs to be overridden, because this is the object that is
eventually executing the plugin user func (Extbase).

This override is minimal though, as only the render function is
overridden and the parent method is called as well after the
request is modified/ restored.

Fixes #14