[TASK] Move rendering code to page type
In order to make transition to TYPO3 v13 more easy code wise
the rendering is now done without hacks by transitioning the
rendering to a regular page type. When a Topwire Context
is detected, the page type is switched to the internal Topwire
page type. The rendering middleware now only does consistency checks
and the alternative TypoScript with the page type definition
is provided statically.

This should be non breaking.
The only side effect is, that default page config is now not
picked up by Topwire. The impact of this should be low to non existent.