[BUGFIX] Make title propagation work with Turbo 8
The title propagation feature implementation was done
because Turbo in older versions did NOT do a full page transition
for frames even with action="advance" and the title tag
being present in the head section of a returned document.

Turbo 8 now does exactly that which is in general good news,
because it works as expected, but bad news for our current
implementation as this broke, when only a part of the HTML document,
the frame, is in the HTML response. Because now an existing title tag
in the head section is discarded, making our JS code fail.

To fix this, our own implementation is completely removed
in favour of returning a full HTML document now, when the
frame requests a full document.